+++ lib/liquex/representable.ex
@@ -4,21 +4,21 @@ defprotocol Liquex.Representable do
   @spec represent(term, boolean) :: term
   def represent(representable, lazy \\ false)
 
-  @spec is_lazy(term) :: boolean
-  def is_lazy(representable)
+  @spec lazy?(term) :: boolean
+  def lazy?(representable)
 end
 
 defimpl Liquex.Representable, for: [Date, Time, DateTime, NaiveDateTime] do
   def represent(representable, _), do: representable
-  def is_lazy(_), do: false
+  def lazy?(_), do: false
 end
 
 defimpl Liquex.Representable, for: Any do
   alias Liquex.Represent
 
   def represent(representable, lazy), do: Represent.represent(representable, lazy)
-  def is_lazy(representable) when is_list(representable), do: true
-  def is_lazy(representable) when is_map(representable), do: true
+  def lazy?(representable) when is_list(representable), do: true
+  def lazy?(representable) when is_map(representable), do: true
 
-  def is_lazy(_), do: false
+  def lazy?(_), do: false
 end